refactor business data layer packages
