+++ app/services/sales-api/handlers/v1/usergrp/usergrp.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/ardanlabs/service/business/core/user"
 	"github.com/ardanlabs/service/business/cview/user/summary"
-	"github.com/ardanlabs/service/business/data/database"
+	"github.com/ardanlabs/service/business/data/tran"
 	"github.com/ardanlabs/service/business/sys/validate"
 	"github.com/ardanlabs/service/business/web/auth"
 	v1 "github.com/ardanlabs/service/business/web/v1"
@@ -39,7 +39,7 @@ func New(user *user.Core, summary *summary.Core, auth *auth.Auth) *Handlers {
 // executeUnderTransaction constructs a new Handlers value with the core apis
 // using a store transaction that was created via middleware.
 func (h *Handlers) executeUnderTransaction(ctx context.Context) (*Handlers, error) {
-	if tx, ok := database.GetTransaction(ctx); ok {
+	if tx, ok := tran.GetTransaction(ctx); ok {
 		user, err := h.user.ExecuteUnderTransaction(tx)
 		if err != nil {
 			return nil, err
