+++ app/tooling/sales-admin/commands/migrate.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"time"
 
-	db "github.com/ardanlabs/service/business/data/database/pgx"
 	"github.com/ardanlabs/service/business/data/dbmigrate"
+	db "github.com/ardanlabs/service/business/data/dbsql/pgx"
 )
 
 // ErrHelp provides context that help was given.
