+++ business/core/product/product_test.go
@@ -11,9 +11,9 @@ import (
 
 	"github.com/ardanlabs/service/business/core/product"
 	"github.com/ardanlabs/service/business/core/user"
-	"github.com/ardanlabs/service/business/data/database"
-	db "github.com/ardanlabs/service/business/data/database/pgx"
+	db "github.com/ardanlabs/service/business/data/dbsql/pgx"
 	"github.com/ardanlabs/service/business/data/dbtest"
+	"github.com/ardanlabs/service/business/data/tran"
 	"github.com/ardanlabs/service/foundation/docker"
 	"github.com/google/go-cmp/cmp"
 )
@@ -309,7 +309,7 @@ func transaction(t *testing.T) {
 	// -------------------------------------------------------------------------
 	// Execute under a transaction with rollback
 
-	tran := func(tx database.Transaction) error {
+	f := func(tx tran.Transaction) error {
 		usrCore, err := api.User.ExecuteUnderTransaction(tx)
 		if err != nil {
 			t.Fatalf("Should be able to create new user core: %s.", err)
@@ -354,7 +354,7 @@ func transaction(t *testing.T) {
 		return nil
 	}
 
-	err := database.ExecuteUnderTransaction(ctx, test.Log, db.NewBeginner(test.DB), tran)
+	err := tran.ExecuteUnderTransaction(ctx, test.Log, db.NewBeginner(test.DB), f)
 	if !errors.Is(err, product.ErrInvalidCost) {
 		t.Fatalf("Should NOT be able to add product : %s.", err)
 	}
@@ -387,7 +387,7 @@ func transaction(t *testing.T) {
 	// -------------------------------------------------------------------------
 	// Good transaction
 
-	tran = func(tx database.Transaction) error {
+	f = func(tx tran.Transaction) error {
 		usrCore, err := api.User.ExecuteUnderTransaction(tx)
 		if err != nil {
 			t.Fatalf("Should be able to create new user core: %s.", err)
@@ -432,7 +432,7 @@ func transaction(t *testing.T) {
 		return nil
 	}
 
-	err = database.ExecuteUnderTransaction(ctx, test.Log, db.NewBeginner(test.DB), tran)
+	err = tran.ExecuteUnderTransaction(ctx, test.Log, db.NewBeginner(test.DB), f)
 	if errors.Is(err, product.ErrInvalidCost) {
 		t.Fatalf("Should be able to add product : %s.", err)
 	}
