+++ business/core/user/stores/userdb/model.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/ardanlabs/service/business/core/user"
-	"github.com/ardanlabs/service/business/data/database/pgx/dbarray"
+	"github.com/ardanlabs/service/business/data/dbsql/pgx/dbarray"
 	"github.com/google/uuid"
 )
