+++ business/core/user/user.go
@@ -11,8 +11,8 @@ import (
 	"time"
 
 	"github.com/ardanlabs/service/business/core/event"
-	"github.com/ardanlabs/service/business/data/database"
 	"github.com/ardanlabs/service/business/data/order"
+	"github.com/ardanlabs/service/business/data/tran"
 	"github.com/ardanlabs/service/business/sys/logger"
 	"github.com/google/uuid"
 	"golang.org/x/crypto/bcrypt"
@@ -30,7 +30,7 @@ var (
 // Storer interface declares the behavior this package needs to perists and
 // retrieve data.
 type Storer interface {
-	ExecuteUnderTransaction(tx database.Transaction) (Storer, error)
+	ExecuteUnderTransaction(tx tran.Transaction) (Storer, error)
 	Create(ctx context.Context, usr User) error
 	Update(ctx context.Context, usr User) error
 	Delete(ctx context.Context, usr User) error
@@ -61,7 +61,7 @@ func NewCore(log *logger.Logger, evnCore *event.Core, storer Storer) *Core {
 
 // ExecuteUnderTransaction constructs a new Core value that will use the
 // specified transaction in any store related calls.
-func (c *Core) ExecuteUnderTransaction(tx database.Transaction) (*Core, error) {
+func (c *Core) ExecuteUnderTransaction(tx tran.Transaction) (*Core, error) {
 	trS, err := c.storer.ExecuteUnderTransaction(tx)
 	if err != nil {
 		return nil, err
