+++ business/data/dbsql/pgx/tran.go
@@ -3,7 +3,7 @@ package db
 import (
 	"fmt"
 
-	"github.com/ardanlabs/service/business/data/database"
+	"github.com/ardanlabs/service/business/data/tran"
 	"github.com/jmoiron/sqlx"
 )
 
@@ -14,7 +14,7 @@ type dbBeginner struct {
 
 // NewBeginner constructs a value that implements the database
 // beginner interface.
-func NewBeginner(sqlxDB *sqlx.DB) database.Beginner {
+func NewBeginner(sqlxDB *sqlx.DB) tran.Beginner {
 	return &dbBeginner{
 		sqlxDB: sqlxDB,
 	}
@@ -22,13 +22,13 @@ func NewBeginner(sqlxDB *sqlx.DB) database.Beginner {
 
 // Begin start a transaction and returns a value that implements
 // the core transactor interface.
-func (db *dbBeginner) Begin() (database.Transaction, error) {
+func (db *dbBeginner) Begin() (tran.Transaction, error) {
 	return db.sqlxDB.Beginx()
 }
 
 // GetExtContext is a helper function that extracts the sqlx value
 // from the core transactor interface for transactional use.
-func GetExtContext(tx database.Transaction) (sqlx.ExtContext, error) {
+func GetExtContext(tx tran.Transaction) (sqlx.ExtContext, error) {
 	ec, ok := tx.(sqlx.ExtContext)
 	if !ok {
 		return nil, fmt.Errorf("Transactor(%T) not of a type *sql.Tx", tx)
